feat : LocalStorage, Cookie 클래스 -> Token 클래스로 통합
